🎨(settings) move configurations mixins in settings
Make settings more explicit and prevent cyclic module importation by
moving base configurations mixins to a core settings file.
