--- sandbox/configurations_mixins/rest_framework.py
@@ -1,17 +1,0 @@
-"""
-Django Rest Framework settings mixin
-"""
-
-
-class DRFMixin(object):
-    """
-    Django Rest Framework configuration mixin.
-    NB: DRF picks its settings from the REST_FRAMEWORK namespace on the settings, hence
-    the nesting of all our values inside that prop
-    """
-
-    REST_FRAMEWORK = {
-        "ALLOWED_VERSIONS": ("1.0",),
-        "DEFAULT_VERSION": "1.0",
-        "DEFAULT_VERSIONING_CLASS": "rest_framework.versioning.URLPathVersioning",
-    }
