+++ sandbox/settings.py
@@ -8,9 +8,7 @@ from django.utils.translation import ugettext_lazy as _
 
 import raven
 from configurations import Configuration, values
-
-from configurations_mixins.elasticsearch import ElasticSearchMixin
-from configurations_mixins.rest_framework import DRFMixin
+from elasticsearch import Elasticsearch
 
 BASE_DIR = os.path.dirname(os.path.abspath(__file__))
 DATA_DIR = os.path.join("/", "data")
@@ -47,6 +45,47 @@ def get_release():
 
     # Default: not available
     return "NA"
+
+
+class ElasticSearchMixin(object):
+    """
+    Elastic Search configuration mixin
+
+    You may want to override default configuration by setting the following environment
+    variable:
+
+    * ES_CLIENT
+    """
+
+    ES_CLIENT = Elasticsearch(
+        [values.Value("localhost", environ_name="ES_CLIENT", environ_prefix=None)]
+    )
+    ES_CHUNK_SIZE = 500
+    ES_INDEXES = [
+        "apps.search.indexers.courses.CoursesIndexer",
+        "apps.search.indexers.organizations.OrganizationsIndexer",
+        "apps.search.indexers.subjects.SubjectsIndexer",
+    ]
+
+    ES_DEFAULT_PAGE_SIZE = 10
+
+    COURSE_API_ENDPOINT = "https://www.fun-mooc.fr/fun/api/courses"
+    ORGANIZATION_API_ENDPOINT = "https://www.fun-mooc.fr/fun/api/universities"
+    SUBJECT_API_ENDPOINT = "https://www.fun-mooc.fr/fun/api/course_subjects"
+
+
+class DRFMixin(object):
+    """
+    Django Rest Framework configuration mixin.
+    NB: DRF picks its settings from the REST_FRAMEWORK namespace on the settings, hence
+    the nesting of all our values inside that prop
+    """
+
+    REST_FRAMEWORK = {
+        "ALLOWED_VERSIONS": ("1.0",),
+        "DEFAULT_VERSION": "1.0",
+        "DEFAULT_VERSIONING_CLASS": "rest_framework.versioning.URLPathVersioning",
+    }
 
 
 class Base(DRFMixin, ElasticSearchMixin, Configuration):
